requirements.txt
aiohttp>=3.9.0
aiofiles>=23.0.0
pydub>=0.25.1
pedalboard>=0.9.0
google-genai>=1.0.0
Pillow>=10.0.0
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
# YouTube API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
